[DEV-6316] Move ensure_root_team() to internal.team module
